refactor: move DeploymentDropHelper and DeploymentRecoveryHandler to client.recovery package
